refactor: extract logging service errors
