+++ crates/mesh-llm-host-runtime/src/logging/service.rs
@@ -30,6 +30,8 @@ pub use operational_audit::{
     OperationalAuditSeverity, OperationalAuditSubjectKind,
 };
 
+pub use super::service_errors::BusEnqueueError;
+
 pub use super::bus::{BusEntry, ReplayBus};
 use super::lifecycle::LifecycleRecorder;
 pub use super::lifecycle::{DuplicateTerminalError, LifecycleGuard, TerminalOutcome};
@@ -1982,20 +1984,3 @@ impl LoggingService {
             .expect("worker handle lock starts healthy")
     }
 }
-
-/// Error type returned when bus enqueue fails (shouldn't happen with drop-oldest, but kept for API completeness).
-#[derive(Clone, Debug)]
-pub enum BusEnqueueError {
-    /// The sink is unavailable and the error-audit fallback also failed.
-    SinkUnavailable(String),
-}
-
-impl std::fmt::Display for BusEnqueueError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::SinkUnavailable(msg) => write!(f, "sink unavailable: {}", msg),
-        }
-    }
-}
-
-impl std::error::Error for BusEnqueueError {}
